refactor(presets): simplify jest presets structure
