--- src/presets/index.spec.ts
@@ -1,10 +1,0 @@
-import presets from './index'
-
-describe('presets', () => {
-  test.each([presets.defaults, presets.jsWithTs, presets.jsWithBabel])(
-    'should return correct preset data',
-    (presetData) => {
-      expect(presetData).toMatchSnapshot()
-    },
-  )
-})
